backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
pytest-django
pytest
factory-boy
pytest-factoryboy
django-environ
gunicorn
requests
black
flake8
Pillow
django-cors-headers
